refactor(lotto money): 로또 금액 객체 생성
로또 번호 객체를 따로 생성했다.
Validator에 분산하지 않고 객체가 직접 validate를 하게끔 했다.
