+++ src/main/java/lotto/Application.java
@@ -5,7 +5,7 @@ import ui.IOView;
 public class Application {
     public static void main(String[] args) {
 
-        int money = IOView.getMoney();
+        LottoMoney money = IOView.getMoney();
         LottoTickets lottoTickets = LottoTickets.buy(money);
         IOView.printLottoNumbers(lottoTickets);
